move all things eslint into eslint-plugin
