+++ package.json
@@ -48,6 +48,7 @@
     "@types/node": "^12"
   },
   "resolutions": {
+    "eslint": "7.8.1",
     "lodash": "4.17.19",
     "@definitelytyped/typescript-versions": "^0.0.40",
     "yargs-parser": "^13.1.2"
