+++ packages/bundler/package.json
@@ -33,7 +33,6 @@
     "@rollup/plugin-node-resolve": "^8.1.0",
     "@types/eslint": "^7.2.2",
     "effection": "^0.7.0",
-    "eslint": "^7.8.1",
     "express": "^4.17.1",
     "rollup": "^2.28.0",
     "rollup-plugin-inject-process-env": "^1.3.0"
