+++ packages/bundler/src/index.ts
@@ -7,7 +7,5 @@ export {
   ValidationWarning,
   BundlerState,
   BundlerTypes,
-  Validator
 } from './types';
-export * from './validators';
 export * from './assertions/bundler-assertions';
