--- packages/bundler/src/validators/index.ts
@@ -1,1 +1,0 @@
-export { EslintValidator, EslintValidatorOptions } from './eslint-validator';
