+++ packages/eslint-plugin/src/eslint-validator/eslint-options.ts
