+++ packages/eslint-plugin/src/types.ts
@@ -1,4 +1,10 @@
+export interface Validator<R, O extends Record<string, unknown> = {}> {
+  validate(files: string | readonly string[], options?: O): Promise<R>;
+}
+
 export type ValidationException = {
+  name: string;
+  fileName: string;
   code?: string;
   frame?: string;
   message: string;
@@ -14,3 +20,7 @@ export type ValidationWarning = ValidationException;
 export type ValidationError = ValidationException & {
   stack?: string;
 }
+
+export type ValidationState = 
+| { type: 'INVALID'; errors: ValidationError[]; warnings: ValidationWarning[] }
+| { type: 'VALID'; warnings: ValidationWarning[] }
